pyproject.toml
[tool.poetry]
name = "specs-extractor"
version = "0.1"
description = "Extracts specs from the doc"
authors = ["Lalith Sagar  Devagudi <lalithsagardevagudi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
marshmallow-dataclass = "^8.7.1"
omegaconf = "^2.3.0"
loguru = "^0.7.2"
fastapi = "^0.115.0"
qdrant-client = "^1.11.3"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"
pymupdf4llm = "^0.0.17"
pathlib = "^1.0.1"
pymupdf = "^1.24.11"
arize-phoenix = "^5.2.2"
langchain-cli = "0.0.31"
langchain-openai = "^0.2.8"
pypdf2 = "^3.0.1"
pdf2image = "^1.17.0"
pdfplumber = "^0.11.4"
ironpdf = "^2024.9.1.3"
tabula-py = "^2.10.0"
tabulate = "^0.9.0"
jpype1 = "^1.5.0"
openpyxl = "^3.1.5"
split-image = "^2.0.1"
langchain = "^0.3.15"
langchain-qdrant = "^0.2.0"
langchain-community = "^0.3.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
